ObsCovid/requirements.txt
Django==2.2.14
django-grappelli==2.13.4
django-jazzmin==2.2.5
django-leaflet==0.27.0
django-userforeignkey==0.3.0
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
